Reorganize some files and betterments to pylint
