+++ .coveragerc
@@ -3,8 +3,8 @@ source = src
 omit =
   src/tests/**,
   src/index.py,
-  src/update_mv_targets.py,
-  src/fetch_from_museovirasto.py,
+  src/util/update_mv_targets.py,
+  src/util/fetch_from_museovirasto.py,
   src/util/email_sender.py,
   src/apis/**
