+++ .github/workflows/updater.yml
@@ -20,4 +20,4 @@ jobs:
       - name: Install backend dependencies
         run: pip install -r ./requirements.txt
       - name: Update targets from Museovirasto interface
-        run: python3 src/update_mv_targets.py
+        run: python3 src/util/update_mv_targets.py
