+++ .pylintrc
@@ -20,7 +20,7 @@ fail-on=
 fail-under=10.0
 
 # Files or directories to be skipped. They should be base names, not paths.
-ignore=CVS,tests
+ignore=CVS,tests, src/tests
 
 # Add files or directories matching the regex patterns to the ignore-list. The
 # regex matches against paths.
