+++ src/email_services/weekly_emailer.py
@@ -1,13 +1,14 @@
-# pylint: disable=unused-import, ungrouped-imports
 import datetime
 import sys
 from email.mime.text import MIMEText
 from email.mime.multipart import MIMEMultipart
-from util.config import SENDER_EMAIL, RECEIVER_EMAIL
 from models.dive import Dive
 from models.targetnote import Targetnote
-import mongo
+from util.config import SENDER_EMAIL, RECEIVER_EMAIL
+from util.mongo import connect_to_db
 from util.email_sender import sender
+
+connect_to_db()
 
 
 class Emailer:
