+++ src/models/targetnote.py
@@ -34,10 +34,10 @@ class Targetnote(MongoModel):
             'target': self.target.to_json(),
             'miscellaneous': self.miscellaneous
         }
+
     def to_json_admin(self):
         return {
             'id': str(self._id),
-            # pylint: disable=W0212
             'target_id': str(self.target.target_id),
             # pylint: disable=W0212
             'user_id': str(self.diver._id),
