+++ src/util/mongo.py
@@ -2,4 +2,5 @@ from pymodm.connection import connect
 from util.config import MONGO_URI
 
 
-connect(MONGO_URI, alias='app')
+def connect_to_db():
+    connect(MONGO_URI, alias='app')
